% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbc_surf.R
\name{lbc_surf}
\alias{lbc_surf}
\title{scrape all real estate advertisementsreal estate advertisements
of a specific research in leboincoin.fr}
\usage{
lbc_surf(data.df)
}
\arguments{
\item{a}{data.frame from `scrape_immo()`}
}
\value{
a vector composed with NA and price when price is detecte
}
\description{
This function take parameter to scrape all advertisements of a specifique page
}
\examples{
lbc_surf(df.from.scrape_immo)
}
