% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.lbc.R
\name{region.lbc}
\alias{region.lbc}
\title{Take region avalable list}
\usage{
region.lbc()
}
\value{
a charactere vector within frensh regions from leboncoin
}
\description{
This function tacke the home page from 'leboncoin.fr'
en return a vector with all frensh region
}
\examples{
my.region <- region.lbc()[7]
}
