% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_immo.R
\name{scape_immo}
\alias{scape_immo}
\title{scrape all real estate advertisementsreal estate advertisements
of a specific research in leboincoin.fr}
\usage{
scape_immo(region, page, querry, localisation, prix_max, nb_pieces_min)
}
\arguments{
\item{region}{a string from `region.lbc()`}

\item{page}{on witch page the search is run}

\item{querry}{specified research}

\item{prix_max}{a number from 1 to n corresponding to the order of price bunton}

\item{location}{a string with the specific commune in france}

\item{nb_pieces_mini}{the minimum number of room you are looking for}
}
\value{
a charactere vector within frensh regions from leboncoin
}
\description{
This function take parameter to scrape all advertisements of a specifique page
}
\examples{
scape_immo(my.region,1, "Montpellier", 5, 3)
}
